# server/docs/openapi.yaml
openapi: 3.0.0
info:
  title: API do Sistema Financeiro
  description: |
    API completa para gerenciamento de finanças pessoais e empresariais.
    
    ## Funcionalidades Principais
    
    ### Gestão Financeira
    - **Contas**: Gerenciamento de contas bancárias e carteiras
    - **Transações**: Registro de receitas e despesas
    - **Categorias**: Organização de transações por categorias
    - **Relatórios**: Análises e relatórios financeiros
    
    ### Gestão de Clientes e Fornecedores
    - **Clientes**: Cadastro e gestão de clientes
    - **Fornecedores**: Controle de fornecedores
    - **Recebíveis**: Controle de contas a receber
    - **Pagáveis**: Controle de contas a pagar
    
    ### Investimentos e Financiamentos
    - **Investimentos**: Gestão de portfólio de investimentos
    - **Metas de Investimento**: Definição e acompanhamento de metas
    - **Financiamentos**: Controle de empréstimos e financiamentos
    - **Pagamentos**: Gestão de parcelas e pagamentos
    
    ### Recursos Avançados
    - **Autenticação JWT**: Sistema seguro de autenticação
    - **Notificações**: Sistema de alertas e lembretes
    - **Jobs**: Processamento em background
    - **Validações**: Validação robusta de dados
    
    ## Status dos Testes
    - **13/17 suítes de integração** funcionando 100%
    - **142/215 testes** passando (66%)
    - **Execução sequencial** implementada para evitar conflitos
    - **Cobertura de código** configurada
    
    ## Documentação Relacionada
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido através do endpoint de login
...
security:
  - bearerAuth: []
...
paths:
...
  /api/payments/receivables/{receivable_id}/payments:
    post:
      tags:
        - Payments
      summary: Criar pagamento para recebível
      security:
        - bearerAuth: []
      parameters:
        - name: receivable_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Pagamento criado
    get:
      tags:
        - Payments
      summary: Listar pagamentos de recebível
      security:
        - bearerAuth: []
      parameters:
        - name: receivable_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de pagamentos
  /api/payments/payables/{payable_id}/payments:
    post:
      tags:
        - Payments
      summary: Criar pagamento para pagável
      security:
        - bearerAuth: []
      parameters:
        - name: payable_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Pagamento criado
    get:
      tags:
        - Payments
      summary: Listar pagamentos de pagável
      security:
        - bearerAuth: []
      parameters:
        - name: payable_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de pagamentos
  /api/payments/{id}:
    delete:
      tags:
        - Payments
      summary: Excluir pagamento
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pagamento removido
...
  /api/fixed-accounts/{id}/pay:
    post:
      tags:
        - FixedAccounts
      summary: Marcar conta fixa como paga
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Pagamento registrado
...
  /api/receivables/{id}/payments:
    get:
      tags:
        - Receivables
      summary: Pagamentos de recebível
      description: Lista pagamentos de uma conta a receber
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de pagamentos
    post:
      tags:
        - Receivables
      summary: Registrar pagamento
      description: Registra um pagamento para uma conta a receber
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Pagamento registrado

  # Endpoints específicos de Pagáveis
...
  /api/payables/{id}/payments:
    get:
      tags:
        - Payables
      summary: Pagamentos de pagável
      description: Lista pagamentos de uma conta a pagar
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de pagamentos
    post:
      tags:
        - Payables
      summary: Adicionar pagamento
      description: Adiciona um pagamento a uma conta a pagar
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Pagamento adicionado

  # Endpoints específicos de Clientes
...
  /api/financings:
    get:
      tags:
        - Financings
      summary: Listar financiamentos
      description: Lista todos os financiamentos do usuário com filtros e paginação
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Número da página
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Limite de itens por página
        - name: financing_type
          in: query
          schema:
            type: string
            enum: [hipoteca, emprestimo_pessoal, veiculo, outros]
          description: Filtrar por tipo de financiamento
        - name: creditor_id
          in: query
          schema:
            type: integer
          description: Filtrar por credor
        - name: status
          in: query
          schema:
            type: string
            enum: [ativo, quitado, inadimplente]
          description: Filtrar por status
        - name: amortization_method
          in: query
          schema:
            type: string
            enum: [SAC, Price]
          description: Filtrar por método de amortização
        - name: start_date_from
          in: query
          schema:
            type: string
            format: date
          description: Data de início a partir de
        - name: start_date_to
          in: query
          schema:
            type: string
            format: date
          description: Data de início até
      responses:
        '200':
          description: Lista de financiamentos
    post:
      tags:
        - Financings
      summary: Criar financiamento
      description: Cria um novo financiamento
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                creditor_id:
                  type: integer
                  description: ID do credor
                financing_type:
                  type: string
                  enum: [hipoteca, emprestimo_pessoal, veiculo, outros]
                  description: Tipo de financiamento
                total_amount:
                  type: number
                  minimum: 0.01
                  description: Valor total financiado
                interest_rate:
                  type: number
                  minimum: 0
                  description: Taxa de juros anual (decimal)
                term_months:
                  type: integer
                  minimum: 1
                  description: Prazo em meses
                start_date:
                  type: string
                  format: date
                  description: Data de início do financiamento
                description:
                  type: string
                  description: Descrição do financiamento
                contract_number:
                  type: string
                  description: Número do contrato
                payment_method:
                  type: string
                  enum: [boleto, debito_automatico, cartao, pix]
                  description: Método de pagamento
                observations:
                  type: string
                  description: Observações adicionais
                amortization_method:
                  type: string
                  enum: [SAC, Price]
                  description: Método de amortização
              required:
                - creditor_id
                - financing_type
                - total_amount
                - interest_rate
                - term_months
                - start_date
                - amortization_method
      responses:
        '201':
          description: Financiamento criado com sucesso
...
  /api/financings/{id}:
    get:
      tags:
        - Financings
      summary: Obter financiamento
      description: Obtém um financiamento específico
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Dados do financiamento
    put:
      tags:
        - Financings
      summary: Atualizar financiamento
      description: Atualiza um financiamento existente
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Financiamento atualizado
    delete:
      tags:
        - Financings
      summary: Excluir financiamento
      description: Remove um financiamento
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Financiamento removido
...
  /api/financing-payments:
    get:
      tags:
        - FinancingPayments
      summary: Listar pagamentos de financiamentos
      description: Lista todos os pagamentos do usuário com filtros e paginação
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Número da página
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Limite de itens por página
        - name: financing_id
          in: query
          schema:
            type: integer
          description: Filtrar por financiamento
        - name: account_id
          in: query
          schema:
            type: integer
          description: Filtrar por conta
        - name: payment_method
          in: query
          schema:
            type: string
            enum: [boleto, debito_automatico, cartao, pix, transferencia]
          description: Filtrar por método de pagamento
        - name: payment_type
          in: query
          schema:
            type: string
            enum: [parcela, parcial, antecipado]
          description: Filtrar por tipo de pagamento
        - name: status
          in: query
          schema:
            type: string
            enum: [pago, pendente, atrasado]
          description: Filtrar por status
        - name: payment_date_from
          in: query
          schema:
            type: string
            format: date
          description: Data de pagamento a partir de
        - name: payment_date_to
          in: query
          schema:
            type: string
            format: date
          description: Data de pagamento até
      responses:
        '200':
          description: Lista de pagamentos
    post:
      tags:
        - FinancingPayments
      summary: Criar pagamento de financiamento
      description: Cria um novo pagamento de financiamento
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                financing_id:
                  type: integer
                  description: ID do financiamento
                account_id:
                  type: integer
                  description: ID da conta
                installment_number:
                  type: integer
                  minimum: 1
                  description: Número da parcela
                payment_amount:
                  type: number
                  minimum: 0.01
                  description: Valor total do pagamento
                principal_amount:
                  type: number
                  minimum: 0
                  description: Valor da amortização
                interest_amount:
                  type: number
                  minimum: 0
                  description: Valor dos juros
                payment_date:
                  type: string
                  format: date
                  description: Data do pagamento
                payment_method:
                  type: string
                  enum: [boleto, debito_automatico, cartao, pix, transferencia]
                  description: Método de pagamento
                payment_type:
                  type: string
                  enum: [parcela, parcial, antecipado]
                  default: parcela
                  description: Tipo de pagamento
                observations:
                  type: string
                  description: Observações do pagamento
              required:
                - financing_id
                - account_id
                - installment_number
                - payment_amount
                - principal_amount
                - interest_amount
                - payment_date
                - payment_method
      responses:
        '201':
          description: Pagamento registrado com sucesso
  /api/financing-payments/{id}:
    get:
      tags:
        - FinancingPayments
      summary: Obter pagamento específico
      description: Obtém um pagamento específico
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Dados do pagamento
    put:
      tags:
        - FinancingPayments
      summary: Atualizar pagamento
      description: Atualiza um pagamento existente
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pagamento atualizado
    delete:
      tags:
        - FinancingPayments
      summary: Excluir pagamento
      description: Remove um pagamento
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pagamento removido
  /api/financing-payments/{id}/pay:
    post:
      tags:
        - FinancingPayments
      summary: Pagar parcela
      description: Marca uma parcela como paga
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Parcela marcada como paga
...